A quick'n'dirty test to check that running code from external SRAM works.
Copy a simple blink function into the SRAM on the FSMC and call it to see
that everything works.

The blink is visibly quite a bit slower, due to external SRAM being
slower than internal flash accelerator.

This only works because the blink function happens to be compiled by
GCC in a way that does not contain any absolute branches or references
any other code that ends up incorrectly after copying the code.

For some reason the low bit of the function address to be called needs
to be set, some ARM think presumably.